outputs(wavefront): reduce code duplication
