+++ plugins/serializers/wavefront/wavefront.go
@@ -3,7 +3,6 @@ package wavefront
 import (
 	"log"
 	"strconv"
-	"strings"
 	"sync"
 
 	"github.com/influxdata/telegraf"
@@ -18,27 +17,6 @@ type WavefrontSerializer struct {
 	scratch                  buffer
 	mu                       sync.Mutex // buffer mutex
 }
-
-// catch many of the invalid chars that could appear in a metric or tag name
-var sanitizedChars = strings.NewReplacer(
-	"!", "-", "@", "-", "#", "-", "$", "-", "%", "-", "^", "-", "&", "-",
-	"*", "-", "(", "-", ")", "-", "+", "-", "`", "-", "'", "-", "\"", "-",
-	"[", "-", "]", "-", "{", "-", "}", "-", ":", "-", ";", "-", "<", "-",
-	">", "-", ",", "-", "?", "-", "/", "-", "\\", "-", "|", "-", " ", "-",
-	"=", "-",
-)
-
-// catch many of the invalid chars that could appear in a metric or tag name
-var strictSanitizedChars = strings.NewReplacer(
-	"!", "-", "@", "-", "#", "-", "$", "-", "%", "-", "^", "-", "&", "-",
-	"*", "-", "(", "-", ")", "-", "+", "-", "`", "-", "'", "-", "\"", "-",
-	"[", "-", "]", "-", "{", "-", "}", "-", ":", "-", ";", "-", "<", "-",
-	">", "-", "?", "-", "\\", "-", "|", "-", " ", "-", "=", "-",
-)
-
-var tagValueReplacer = strings.NewReplacer("\"", "\\\"", "*", "-")
-
-var pathReplacer = strings.NewReplacer("_", ".")
 
 type MetricPoint struct {
 	Metric    string
@@ -70,11 +48,7 @@ func (s *WavefrontSerializer) serializeMetric(m telegraf.Metric) {
 			name = s.Prefix + m.Name() + metricSeparator + fieldName
 		}
 
-		if s.UseStrict {
-			name = strictSanitizedChars.Replace(name)
-		} else {
-			name = sanitizedChars.Replace(name)
-		}
+		name = Sanitize(s.UseStrict, name)
 
 		if !s.DisablePrefixConversions {
 			name = pathReplacer.Replace(name)
@@ -182,11 +156,7 @@ func formatMetricPoint(b *buffer, metricPoint *MetricPoint, s *WavefrontSerializ
 
 	for k, v := range metricPoint.Tags {
 		b.WriteString(` "`)
-		if s.UseStrict {
-			b.WriteString(strictSanitizedChars.Replace(k))
-		} else {
-			b.WriteString(sanitizedChars.Replace(k))
-		}
+		b.WriteString(Sanitize(s.UseStrict, k))
 		b.WriteString(`"="`)
 		b.WriteString(tagValueReplacer.Replace(v))
 		b.WriteChar('"')
